change the date type  of transaction table

+++ config/schema/20250111004616_transaction.sql
@@ -3,7 +3,7 @@ CREATE TABLE transactions (
   id VARCHAR(64) NOT NULL PRIMARY KEY,
   name VARCHAR(32) NOT NULL,
   date TIMESTAMP NOT NULL,
-  amount INT NOT NULL,
+  amount DECIMAL(10, 2) NOT NULL,
   category_id VARCHAR(64) NOT NULL,
   is_deleted BOOLEAN DEFAULT FALSE,
   created_by VARCHAR(64) DEFAULT NULL,
